♻️ refactor(SCRUM-259): type, 상수 데이터 폴더 이동
